requirements.txt
transformers==4.46.2
pytorchvideo
imageio
argparse
evaluate
opencv-python
Pillow